[minor] move event constants into the mig package
