+++ src/mig/scheduler/agents_queues.go
@@ -9,7 +9,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"mig"
-	"mig/event"
 	"time"
 
 	"github.com/streadway/amqp"
@@ -102,7 +101,7 @@ func getHeartbeats(msg amqp.Delivery, ctx Context) (err error) {
 		desc := fmt.Sprintf("getHeartbeats(): Agent '%s' is not authorized", agt.QueueLoc)
 		ctx.Channels.Log <- mig.Log{Desc: desc}.Warning()
 		// send an event to notify workers of the failed agent auth
-		err = sendEvent(event.Q_Agt_Auth_Fail, msg.Body, ctx)
+		err = sendEvent(mig.Ev_Q_Agt_Auth_Fail, msg.Body, ctx)
 		if err != nil {
 			panic(err)
 		}
@@ -123,9 +122,9 @@ func getHeartbeats(msg amqp.Delivery, ctx Context) (err error) {
 				ctx.Channels.Log <- mig.Log{Desc: fmt.Sprintf("Heartbeat DB insertion failed with error '%v' for agent '%s'", err, agt.Name)}.Err()
 			}
 			// notify the agt.new event queue
-			err = sendEvent(event.Q_Agt_New, msg.Body, ctx)
+			err = sendEvent(mig.Ev_Q_Agt_New, msg.Body, ctx)
 			if err != nil {
-				ctx.Channels.Log <- mig.Log{Desc: fmt.Sprintf("Failed to send migevent to %s: %v", err, event.Q_Agt_New)}.Err()
+				ctx.Channels.Log <- mig.Log{Desc: fmt.Sprintf("Failed to send migevent to %s: %v", err, mig.Ev_Q_Agt_New)}.Err()
 			}
 		} else {
 			// the agent exists in database. reuse the existing ID, and keep the status if it was
